+++ src/kyber_scalar/randombytes.c
@@ -1,10 +1,6 @@
 #include <kyber_scalar/randombytes.h>
-#include <libfam/format.h>
 #include <libfam/rng.h>
-#include <stddef.h>
-#include <stdint.h>
-#include <stdlib.h>
 
-void randombytes(u8 *out, size_t outlen, Rng *rng) {
+void randombytes(u8 *out, u64 outlen, Rng *rng) {
 	rng_gen(rng, out, outlen);
 }
